requirements.txt
opencv-python~=4.5.4.58
numpy~=1.21.4
lxml~=4.6.4
tqdm~=4.62.3
pandas~=1.3.4
torch~=1.10.0
scikit-image~=0.18.3
torchvision~=0.11.1
decouple~=0.0.7
Pillow~=8.4.0